requirements.txt
bottle==0.12.9
bottle-websocket==0.2.9
equests==2.12.3
